Revert "Feat: 포인트 시스템 구현 (조회, 충전, 사용, 히스토리 내역 조회 및 동시성 제어)"
