+++ 1.aboutTDD/package.json
@@ -25,7 +25,6 @@
         "@nestjs/core": "^10.3.2",
         "@nestjs/platform-express": "^10.3.2",
         "@nestjs/terminus": "^10.2.0",
-        "async-mutex": "^0.5.0",
         "class-transformer": "^0.5.1",
         "class-validator": "^0.14.1",
         "raw-body": "^2.5.2",
