+++ 1.aboutTDD/src/database/pointhistory.table.ts
