+++ 1.aboutTDD/src/point/point.controller.ts
@@ -1,0 +1,51 @@
+import { Body, Controller, Get, Param, Patch, ValidationPipe } from '@nestjs/common';
+import { PointHistory, TransactionType, UserPoint } from './point.model';
+import { UserPointTable } from 'src/database/userpoint.table';
+import { PointHistoryTable } from 'src/database/pointhistory.table';
+import { PointBody as PointDto } from './point.dto';
+
+@Controller('/point')
+export class PointController {
+    constructor(
+        private readonly userDb: UserPointTable,
+        private readonly historyDb: PointHistoryTable,
+    ) {}
+
+    /**
+     * TODO - 특정 유저의 포인트를 조회하는 기능을 작성해주세요.
+     */
+    @Get(':id')
+    async point(@Param('id') id): Promise<UserPoint> {
+        const userId = Number.parseInt(id);
+        return { id: userId, point: 0, updateMillis: Date.now() };
+    }
+
+    /**
+     * TODO - 특정 유저의 포인트 충전/이용 내역을 조회하는 기능을 작성해주세요.
+     */
+    @Get(':id/histories')
+    async history(@Param('id') id): Promise<PointHistory[]> {
+        const userId = Number.parseInt(id);
+        return [];
+    }
+
+    /**
+     * TODO - 특정 유저의 포인트를 충전하는 기능을 작성해주세요.
+     */
+    @Patch(':id/charge')
+    async charge(@Param('id') id, @Body(ValidationPipe) pointDto: PointDto): Promise<UserPoint> {
+        const userId = Number.parseInt(id);
+        const amount = pointDto.amount;
+        return { id: userId, point: amount, updateMillis: Date.now() };
+    }
+
+    /**
+     * TODO - 특정 유저의 포인트를 사용하는 기능을 작성해주세요.
+     */
+    @Patch(':id/use')
+    async use(@Param('id') id, @Body(ValidationPipe) pointDto: PointDto): Promise<UserPoint> {
+        const userId = Number.parseInt(id);
+        const amount = pointDto.amount;
+        return { id: userId, point: amount, updateMillis: Date.now() };
+    }
+}
