+++ 1.aboutTDD/src/point/point.dto.ts
@@ -1,0 +1,6 @@
+import { IsInt, isPositive } from 'class-validator';
+
+export class PointBody {
+    @IsInt()
+    amount: number;
+}
