+++ 1.aboutTDD/src/point/point.model.ts
