+++ 1.aboutTDD/src/point/point.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { PointController } from './point.controller';
-import { DatabaseModule } from '../database/database.module';
+import { DatabaseModule } from 'src/database/database.module';
 import { PointService } from './point.service';
 
 @Module({
