--- src/point/point.dto.ts
@@ -1,7 +1,0 @@
-import { IsInt, IsPositive } from 'class-validator';
-
-export class PointBody {
-    @IsInt()
-    @IsPositive()
-    amount: number;
-}
